pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genbank-to-biodataset"
version = "0.1.0"
description = "CLI pipeline to turn GenBank queries into ML-ready datasets."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
dependencies = [
    "requests>=2.31",
    "pyyaml>=6.0",
    "pandas>=2.1",
]

[project.scripts]
genbank2bio = "genbank_to_biodataset.cli:main"
